fix compile error on clang build

+++ src/qwt3d_lighting.cpp
@@ -175,7 +175,7 @@ void ExtGLWidget::applyLight(unsigned light)
   glRotatef( lights_[light].rot.x-90, 1.0, 0.0, 0.0 ); 
   glRotatef( lights_[light].rot.y   , 0.0, 1.0, 0.0 ); 
   glRotatef( lights_[light].rot.z   , 0.0, 0.0, 1.0 );
-  GLfloat lightPos[4] = { lights_[light].shift.x, lights_[light].shift.y, lights_[light].shift.z, 1.0};
+  GLfloat lightPos[4] = { (float)lights_[light].shift.x, (float)lights_[light].shift.y, (float)lights_[light].shift.z, 1.0};
   GLenum le = lightEnum(light);
   glLightfv(le, GL_POSITION, lightPos);  
 }
